Split archetype from dexterity implementation.
